:recycle: refactor: EmailScheduler 리팩토링
- 스레드 풀을 사용하여 이메일 전송 속도 개선

- List 대신 Map 자료구조를 사용하여 이메일 알림에 필요한 데이터를 생성하는 로직을 간소화함

- [할 일] 수신자 이메일 주소 검증하기

- [할 일] 큐 기반 처리 반영하기
